pyproject.toml
[project]
name = "decision-tree"
version = "0.1.0"
description = "Cleaning and analysis scripts for TerraFund portfolio data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asana",
    "boto3",
    "exactextract",
    "fiona",
    "gdal",
    "matplotlib",
    "geopandas",
    "pandas",
    "pytest",
    "pytest-xdist",
    "pyyaml",
    "rasterio",
    "requests",
    "pyproj[data]>=3.6.1",  # Includes PROJ database for proj.db
    "seaborn",
    "shapely",
    "tqdm",
    "utm",
    "tm-api-utils @ git+https://github.com/wri/terramatch-researcher-api.git@develop"
]

[tool.uv.sources]
gdal = [
 { index = "gdal-wheels", marker = "sys_platform == 'linux'" },
 { index = "geospatial_wheels", marker = "sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "geospatial_wheels"
url = "https://nathanjmcdougall.github.io/geospatial-wheels-index/"
explicit = true

[[tool.uv.index]]
name = "gdal-wheels"
url = "https://gitlab.com/api/v4/projects/61637378/packages/pypi/simple"
explicit = true

[tool.poetry.dependencies]
pyproj = { version = "^3.6.1", extras = ["data"] }  # Includes proj.db file

[tool.env]
# Set PROJ_LIB to the directory where pyproj installs its data files
PROJ_DATA = "${VIRTUAL_ENV}/share/proj"
PROJ_LIB  = "${VIRTUAL_ENV}/share/proj"


